// client/src/types/VehicleRepair.ts
export default interface IVehicleRepair {
  documentNo: string;
  vehicleNo: string;
  invoiceNo: string;
  vehicleType: string;
  workshop: string;
  description: string;
  date: string;
  invoiceDate: string;
  repairStartDate: string;
  repairEndDate: string;
  meterReading: string;
  repairType: string;
  repairCost: string;
  location: string;
  remark: string;
}